requirements.txt
fastapi
Pillow
torch
torchvision
python-multipart
uvicorn
fastapi[standard]